+++ generate_reads/src/utils/generate_fragments.rs
@@ -12,14 +12,15 @@ use std::collections::VecDeque;
 use simple_rng::NeatRng;
 use crate::errors::GenerateReadsErrors;
 
-pub fn generate_reads(
+pub fn generate_fragments(
     sequence_length: usize,
     read_length: usize,
+    start: usize,
     coverage: usize,
     paired_ended: bool,
     fragment_model: &FragmentLengthModel,
     rng: &mut NeatRng,
-) -> Result<Vec<(usize, usize, Option<usize>, usize)>, GenerateReadsErrors> {
+) -> Result<Vec<(usize, usize)>, GenerateReadsErrors> {
     // Takes:
     // sequence_length: The length of the sequence to generate reads for.
     // read_length: the length ef the reads for this run
@@ -48,50 +49,27 @@ pub fn generate_reads(
     let fragments: Vec<(usize, usize)> = cover_dataset(
         sequence_length,
         read_length,
+        start,
         fragment_pool,
         coverage,
         rng
-    );
+    )?;
 
-    // TODO Generate single or paired ended reads from the fragments
-    let mut read_set: Vec<(usize, usize, Option<usize>, usize)> = Vec::new();
-    if paired_ended {
-        for fragment in fragments {
-            // Push a paired ended read off the fragment
-            read_set.push((
-                fragment.0, 
-                fragment.0 + read_length, 
-                Some(fragment.1 - read_length), 
-                fragment.1
-            ))
-        }
-    } else {
-        // Push a single ended read
-        for fragment in fragments {
-            read_set.push((
-                fragment.0, 
-                fragment.0 + read_length,
-                None,
-                fragment.1,
-            ))
-        }
-    }
-
-    debug!("Outputting read set");
-    if read_set.is_empty() {
+    if fragments.is_empty() {
         Err(GenerateReadsErrors::GenerateReadsError)
     } else {
-        Ok(read_set)
+        Ok(fragments)
     }
 }
 
 fn cover_dataset(
     span_length: usize,
     read_length: usize,
+    read_start: usize,
     mut fragment_pool: Vec<usize>,
     coverage: usize,
     rng: &mut NeatRng,
-) -> Vec<(usize, usize)> {
+) -> Result<Vec<(usize, usize)>, GenerateReadsErrors> {
     // Takes:
     // span_length: Total number of bases in the sequence
     // read_length: The length of the reads for this run
@@ -117,15 +95,14 @@ fn cover_dataset(
         cover_fragment_pool = VecDeque::from([read_length*2]);
     } else {
         // shuffle the fragment pool
-        rng.shuffle_in_place(&mut fragment_pool)
-            .expect("Generate reads failed while shuffling the fragment pool");
+        rng.shuffle_in_place(&mut fragment_pool)?;
         cover_fragment_pool = VecDeque::from(fragment_pool)
     }
     // Gap size to keep track of how many uncovered bases we have per layer, to help decide if we
     // need more layers
     let mut layer_count: usize = 0;
-    // start this party off at zero.
-    let mut start: usize = 0;
+    // start this party off somewhere random.
+    let mut start = ((rng.rand_int()?) as usize) % (read_length/4);
     // create coverage number of layers
     while layer_count < coverage {
         let fragment_length = cover_fragment_pool.pop_front().unwrap();
@@ -137,7 +114,8 @@ fn cover_dataset(
             layer_count += 1;
             continue
         }
-        fragment_set.push((start, temp_end));
+        // Read start ensures our fragments are within the sequence block
+        fragment_set.push((start+read_start, temp_end+read_start));
         // Picks a number between zero and a quarter of a read length
         let wildcard: usize = (rng.rand_u32().unwrap() % 10) as usize;
         // adds to the start to give it some spice
@@ -148,7 +126,10 @@ fn cover_dataset(
             start = start % span_length;
         }
     }
-    fragment_set
+    fragment_set.sort_by(
+        |a, b| a.0.cmp(&b.0)
+    );
+    Ok(fragment_set)
 }
 
 #[cfg(test)]
@@ -167,7 +148,14 @@ mod tests {
             "Cruel".to_string(),
             "World".to_string(),
         ]).unwrap();
-        let cover = cover_dataset(span_length, read_length, fragment_pool, coverage, &mut rng);
+        let cover = cover_dataset(
+            span_length, 
+            read_length, 
+            0,
+            fragment_pool, 
+            coverage, 
+            &mut rng
+        ).unwrap();
         assert_eq!(cover[0], (0, 10))
     }
 
@@ -182,7 +170,14 @@ mod tests {
             "Cruel".to_string(),
             "World".to_string(),
         ]).unwrap();
-        let cover = cover_dataset(span_length, read_length, fragment_pool, coverage, &mut rng);
+        let cover = cover_dataset(
+            span_length, 
+            read_length,
+            0,
+            fragment_pool, 
+            coverage, 
+            &mut rng
+        ).unwrap();
         assert_eq!(cover[0], (0, 300))
     }
 
@@ -197,15 +192,16 @@ mod tests {
             "World".to_string(),
         ]).unwrap();
         let fragment_model = FragmentLengthModel::default().unwrap();
-        let reads = generate_reads(
+        let reads = generate_fragments(
             2000,
             read_length,
+            0,
             coverage,
             paired_ended,
             &fragment_model,
             &mut rng,
         ).unwrap();
-        assert!(reads.contains(&(0, 10, None, 20)));
+        assert!(reads.contains(&(0, 20)));
     }
 
     #[test]
@@ -222,9 +218,10 @@ mod tests {
             "World".to_string(),
         ]).unwrap();
         let fragment_model = FragmentLengthModel::default().unwrap();
-        let run1 = generate_reads(
+        let run1 = generate_fragments(
             sequnce.len(),
             read_length,
+            0,
             coverage,
             paired_ended,
             &fragment_model,
@@ -232,9 +229,10 @@ mod tests {
         ).unwrap();
 
         let fragment_model = FragmentLengthModel::default().unwrap();
-        let run2 = generate_reads(
+        let run2 = generate_fragments(
             sequnce.len(),
             read_length,
+            0,
             coverage,
             paired_ended,
             &fragment_model,
@@ -256,9 +254,10 @@ mod tests {
             "World".to_string(),
         ]).unwrap();
         let fragment_model = FragmentLengthModel::default().unwrap();
-        let reads = generate_reads(
+        let reads = generate_fragments(
             sequence.len(),
             read_length,
+            0,
             coverage,
             paired_ended,
             &fragment_model,